+++ tests/mediawiki_test.py
@@ -1,8 +1,10 @@
+# -*- coding: utf-8 -*-
 '''
 Unittest class
 '''
-# -*- coding: utf-8 -*-
+
 from __future__ import (unicode_literals, print_function)
+import sys
 import unittest
 import json
 from datetime import timedelta
@@ -27,6 +29,8 @@ class MediaWikiOverloaded(MediaWiki):
             self.requests = json.load(file_handle)
         with open('./tests/mock_responses.json', 'r') as file_handle:
             self.responses = json.load(file_handle)
+        self.tree_path = './tests/mock_categorytree.json'
+
         MediaWiki.__init__(self, url=url, lang=lang, timeout=timeout,
                            rate_limit=rate_limit,
                            rate_limit_wait=rate_limit_wait)
@@ -435,7 +439,7 @@ class TestMediaWikiSummary(unittest.TestCase):
         res = response['sumarize_sent_5']
         sumr = site.summary('chess', sentences=5)
         self.assertEqual(res, sumr)
-        self.assertEqual(len(res), 466)
+        # self.assertEqual(len(res), 466)
 
     def test_page_summary_chars(self):
         ''' test page summarize - chars '''
@@ -455,7 +459,7 @@ class TestMediaWikiSummary(unittest.TestCase):
         pag = site.page('chess')
         sumr = pag.summarize(sentences=5)
         self.assertEqual(res, sumr)
-        self.assertEqual(len(res), 466)
+        # self.assertEqual(len(res), 466)
 
 
 class TestMediaWikiCategoryMembers(unittest.TestCase):
@@ -500,7 +504,7 @@ class TestMediaWikiCategoryMembers(unittest.TestCase):
         ctm = site.categorymembers('Disambiguation categories', results=None)
         self.assertEqual(list(ctm), res)
         self.assertEqual(len(res[0]), 0)
-        self.assertEqual(len(res[1]), 1274)
+        self.assertEqual(len(res[1]), 1290)  # difficult if it changes sizes
 
 
 class TestMediaWikiExceptions(unittest.TestCase):
@@ -666,8 +670,19 @@ class TestMediaWikiExceptions(unittest.TestCase):
         try:
             MediaWikiOverloaded(url=url, lang='fr')
         except MediaWikiAPIURLError as ex:
-            response = site.responses[url]
+            response = site.responses[site.api_url]
             self.assertEqual(ex.message, response['api_url_error_msg'])
+
+    def test_api_url_on_error_reset(self):
+        ''' test api url error resets to original URL '''
+        site = MediaWikiOverloaded()  # something to use to lookup results
+        url = 'http://french.wikipedia.org/w/api.php'
+        wiki = 'http://en.wikipedia.org/w/api.php'
+        try:
+            MediaWikiOverloaded(url=url, lang='fr')
+        except MediaWikiAPIURLError as ex:
+            self.assertNotEqual(site.api_url, url)
+            self.assertEqual(site.api_url, wiki)
 
     def test_http_timeout_msg(self):
         ''' test the http timeout message '''
@@ -839,8 +854,8 @@ class TestMediaWikiPage(unittest.TestCase):
             msg = 'Either a title or a pageid must be specified'
             self.assertEqual(str(ex), msg)
 
-    def test_page_value_err(self):
-        ''' test that ValueError is thrown from random'''
+    def test_page_value_err_none(self):
+        ''' test that ValueError is thrown from None '''
         site = MediaWikiOverloaded()
         self.assertRaises(ValueError, lambda: site.page(None))
 
@@ -953,9 +968,34 @@ class TestMediaWikiPage(unittest.TestCase):
         ''' test page html '''
         self.assertEqual(self.pag.html, self.response['arya']['html'])
 
-    def test_page_repr(self):
-        ''' test page representation '''
+    def test_page_str(self):
+        ''' test page string representation '''
         self.assertEqual(str(self.pag), '''<MediaWikiPage 'Arya Stark'>''')
+
+    def test_page_repr(self):
+        ''' test page repr without unicode '''
+        self.assertEqual(repr(self.pag), '''<MediaWikiPage 'Arya Stark'>''')
+
+    def test_page_unicode(self):
+        ''' test with unicode representation '''
+        site = MediaWikiOverloaded()
+        page = site.page('Jacques Léonard Muller')
+        if sys.version_info < (3, 0):
+            res = unicode(page)
+        else:
+            res = str(page)
+        self.assertEqual(res, '''<MediaWikiPage 'Jacques Léonard Muller'>''')
+
+    def test_page_repr_2(self):
+        ''' test page string representation '''
+        site = MediaWikiOverloaded()
+        page = site.page('Jacques Léonard Muller')
+        name = u'''<MediaWikiPage 'Jacques Léonard Muller'>'''
+        if sys.version_info > (3, 0):
+            res = repr(page)
+        else:
+            res = unicode(page)
+        self.assertEqual(res, name)
 
     def test_page_eq(self):
         ''' test page equality '''
@@ -990,16 +1030,30 @@ class TestMediaWikiPage(unittest.TestCase):
     def test_page_preload(self):
         ''' test preload of page properties '''
         pag = self.site.page('arya', preload=True)
-        self.assertEqual(hasattr(pag, '_content'), True)
-        self.assertEqual(hasattr(pag, '_summary'), True)
-        self.assertEqual(hasattr(pag, '_images'), True)
-        self.assertEqual(hasattr(pag, '_references'), True)
-        self.assertEqual(hasattr(pag, '_links'), True)
-        self.assertEqual(hasattr(pag, '_sections'), True)
-        self.assertEqual(hasattr(pag, '_redirects'), True)
-        self.assertEqual(hasattr(pag, '_coordinates'), True)
-        self.assertEqual(hasattr(pag, '_backlinks'), True)
-        self.assertEqual(hasattr(pag, '_categories'), True)
+        self.assertNotEqual(getattr(pag, '_content'), '')
+        self.assertNotEqual(getattr(pag, '_summary'), False)
+        self.assertNotEqual(getattr(pag, '_images'), False)
+        self.assertNotEqual(getattr(pag, '_references'), False)
+        self.assertNotEqual(getattr(pag, '_links'), False)
+        self.assertNotEqual(getattr(pag, '_sections'), False)
+        self.assertNotEqual(getattr(pag, '_redirects'), False)
+        self.assertNotEqual(getattr(pag, '_coordinates'), False)
+        self.assertNotEqual(getattr(pag, '_backlinks'), False)
+        self.assertNotEqual(getattr(pag, '_categories'), False)
+
+    def test_page_no_preload(self):
+        ''' test page properties that are not set '''
+        pag = self.site.page('arya', preload=False)
+        self.assertEqual(getattr(pag, '_content'), '')
+        self.assertEqual(getattr(pag, '_summary'), False)
+        self.assertEqual(getattr(pag, '_images'), False)
+        self.assertEqual(getattr(pag, '_references'), False)
+        self.assertEqual(getattr(pag, '_links'), False)
+        self.assertEqual(getattr(pag, '_sections'), False)
+        self.assertEqual(getattr(pag, '_redirects'), False)
+        self.assertEqual(getattr(pag, '_coordinates'), False)
+        self.assertEqual(getattr(pag, '_backlinks'), False)
+        self.assertEqual(getattr(pag, '_categories'), False)
 
 
 class TestMediaWikiCategoryTree(unittest.TestCase):
@@ -1008,7 +1062,7 @@ class TestMediaWikiCategoryTree(unittest.TestCase):
     def test_double_category_tree(self):
         ''' test category tree using a list '''
         site = MediaWikiOverloaded()
-        with open('./tests/mock_categorytree.json', 'r') as fpt:
+        with open(site.tree_path, 'r') as fpt:
             res = json.load(fpt)
         cat = site.categorytree(['Chess', 'Ebola'], depth=None)
         self.assertEqual(cat, res)
@@ -1016,7 +1070,7 @@ class TestMediaWikiCategoryTree(unittest.TestCase):
     def test_triple_category_tree_none(self):
         ''' test category tree using a list but one is blank or None '''
         site = MediaWikiOverloaded()
-        with open('./tests/mock_categorytree.json', 'r') as fpt:
+        with open(site.tree_path, 'r') as fpt:
             res = json.load(fpt)
         cat = site.categorytree(['Chess', 'Ebola', None], depth=None)
         self.assertEqual(cat, res)
@@ -1024,7 +1078,7 @@ class TestMediaWikiCategoryTree(unittest.TestCase):
     def test_triple_category_tree_bnk(self):
         ''' test category tree using a list but one is blank or None '''
         site = MediaWikiOverloaded()
-        with open('./tests/mock_categorytree.json', 'r') as fpt:
+        with open(site.tree_path, 'r') as fpt:
             res = json.load(fpt)
         cat = site.categorytree(['Chess', 'Ebola', ''], depth=None)
         self.assertEqual(cat, res)
@@ -1032,7 +1086,7 @@ class TestMediaWikiCategoryTree(unittest.TestCase):
     def test_single_category_tree_list(self):
         ''' test category tree using a list with one element '''
         site = MediaWikiOverloaded()
-        with open('./tests/mock_categorytree.json', 'r') as fpt:
+        with open(site.tree_path, 'r') as fpt:
             res = json.load(fpt)
         cat = site.categorytree(['Chess'], depth=None)
         self.assertEqual(cat['Chess'], res['Chess'])
@@ -1040,7 +1094,7 @@ class TestMediaWikiCategoryTree(unittest.TestCase):
     def test_single_category_tree_str(self):
         ''' test category tree using a string '''
         site = MediaWikiOverloaded()
-        with open('./tests/mock_categorytree.json', 'r') as fpt:
+        with open(site.tree_path, 'r') as fpt:
             res = json.load(fpt)
         cat = site.categorytree('Ebola', depth=None)
         self.assertEqual(cat['Ebola'], res['Ebola'])